requirements.txt
Django>=4.1.6,<4.2
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.9,<2.10
drf-spectacular>=0.27.2,<0.28
